LayoutTests/http/tests/security/contentSecurityPolicy/1.1/scriptnonce-and-scripthash-expected.txt

ALERT: PASS (1/3)
ALERT: PASS (2/3)
ALERT: PASS (3/3)
CONSOLE MESSAGE: line 16: Refused to execute a script because its hash, its nonce, or 'unsafe-inline' does not appear in the script-src directive of the Content Security Policy.
CONSOLE MESSAGE: line 19: Refused to execute a script because its hash, its nonce, or 'unsafe-inline' does not appear in the script-src directive of the Content Security Policy.
This tests the combined use of script hash and script nonce. It passes if two console warnings are visible and the three alerts show PASS.
